refactor: use abs screen coords for boids & MVC class structure
Instead of using the position on canvas (inside viewport), instead we compute
the real position on screen and use those for simulations.
This allows boids to stay on the same absolute coordinates even if the
window is moved or resized.

But the renderer needs to translate every coords before draw by
substracting the left corner of the canvas.