Rmv eslint, jest & non-essentials with crazy dependency tree
Funny enough. Almost any package with a somehow suspicious number of dependencies
was affected by the recent supply chain attack.

Almost like it's part of the JS dev community that is actually responsible
for introducing all these crazy dependencies to their packages over the years,
why most are not even really needed (chalk, really? As if coloring logs is hard).

Almost like exactly what the rest of the industry kept telling them not to do.